+++ Src/Trials/ConsoleApp/webapi/SIS.OpenCore.webapi.csproj
@@ -12,5 +12,9 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="5.3.1" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\BL\SIS.OpenCore.BL.csproj" />
+  </ItemGroup>
+
 
 </Project>
